Exporting charts using the FileServices class
