+++ src/HeatManager.Core/Services/FileServices/IChartExporter.cs
@@ -1,0 +1,8 @@
+
+
+namespace HeatManager.Services.FileServices;
+
+public interface IChartExporter
+{
+    public Task Export(object chartParam);
+}
